# BlockChainAppDev/PetBoarding/docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: 宠物寄养系统API
  description: |
    宠物寄养系统的RESTful API文档，包含用户、宠物、寄养服务、订单、评价和通知等功能。
    
    ## 认证
    大部分API需要通过JWT令牌进行认证，请在请求头中添加`Authorization: Bearer {token}`。
    
    ## 错误处理
    所有API响应均使用统一的格式，错误时会返回相应的错误码和错误信息。
  version: 1.0.0
  contact:
    name: 宠物寄养系统团队
    email: support@petboarding.example.com
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: https://api.petboarding.example.com/api
    description: 生产环境
...
paths:
...
  /pets:
    post:
      tags:
        - 宠物
      summary: 添加宠物
      description: 添加新宠物
      operationId: addPet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetCreateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pets/{petId}:
    get:
      tags:
        - 宠物
      summary: 获取宠物信息
      description: 获取指定宠物的信息
      operationId: getPetInfo
      parameters:
        - name: petId
          in: path
          description: 宠物ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取宠物信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '404':
          description: 宠物不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 宠物
      summary: 更新宠物信息
      description: 更新指定宠物的信息
      operationId: updatePetInfo
      parameters:
        - name: petId
          in: path
          description: 宠物ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetUpdateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 宠物不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 宠物
      summary: 删除宠物
      description: 删除指定宠物
      operationId: deletePet
      parameters:
        - name: petId
          in: path
          description: 宠物ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 宠物不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pets/user/{userId}:
    get:
      tags:
        - 宠物
      summary: 获取用户的所有宠物
      description: 获取指定用户的所有宠物
      operationId: getUserPets
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户的所有宠物
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetListResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 寄养服务API
...
  /boarding/{serviceId}:
    get:
      tags:
        - 寄养服务
      summary: 获取寄养服务信息
      description: 获取指定寄养服务的信息
      operationId: getBoardingServiceInfo
      parameters:
        - name: serviceId
          in: path
          description: 服务ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功获取寄养服务信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardingServiceResponse'
        '404':
          description: 寄养服务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 寄养服务
      summary: 更新寄养服务信息
      description: 更新指定寄养服务的信息
      operationId: updateBoardingServiceInfo
      parameters:
        - name: serviceId
          in: path
          description: 服务ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardingServiceUpdateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardingServiceResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 寄养服务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reviews:
    post:
      tags:
        - 评价
      summary: 创建评价
      description: 创建新评价
      operationId: createReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reviews/{reviewId}:
    get:
      tags:
        - 评价
      summary: 获取评价信息
      description: 获取指定评价的信息
      operationId: getReviewInfo
      parameters:
        - name: reviewId
          in: path
          description: 评价ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功获取评价信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '404':
          description: 评价不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reviews/user/{userId}:
    get:
      tags:
        - 评价
      summary: 获取用户的所有评价
      description: 获取指定用户的所有评价
      operationId: getUserReviews
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: 页码，默认1
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: 每页数量，默认10
          required: false
          schema:
            type: integer
            default: 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功获取用户的所有评价
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewListResponse'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 通知服务API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          nullable: true
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Invalid parameters"
        data:
          type: object
          nullable: true

    # 用户相关
...
    PetCreateRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          example: "Buddy"
        type:
          type: string
          example: "Dog"
        breed:
          type: string
          example: "Golden Retriever"
        age:
          type: integer
          example: 3
        gender:
          type: integer
          enum: [0, 1, 2]
          example: 1
        weight:
          type: number
          format: float
          example: 25.5
        description:
          type: string
          example: "Friendly and energetic"
    PetResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            pet_id:
              type: string
              example: "p123456789"
            user_id:
              type: string
              example: "u123456789"
            name:
              type: string
              example: "Buddy"
            type:
              type: string
              example: "Dog"
            breed:
              type: string
              example: "Golden Retriever"
            age:
              type: integer
              example: 3
            gender:
              type: integer
              example: 1
            weight:
              type: number
              format: float
              example: 25.5
            description:
              type: string
              example: "Friendly and energetic"
            created_at:
              type: string
              format: date-time
              example: "2023-06-01T14:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-06-01T14:00:00Z"
    PetUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: "Buddy Jr"
        age:
          type: integer
          example: 4
        weight:
          type: number
          format: float
          example: 26.2
        description:
          type: string
          example: "Very friendly and energetic"
    PetListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: array
          items:
            type: object
            properties:
              pet_id:
                type: string
                example: "p123456789"
              user_id:
                type: string
                example: "u123456789"
              name:
                type: string
                example: "Buddy"
              type:
                type: string
                example: "Dog"
              breed:
                type: string
                example: "Golden Retriever"
              age:
                type: integer
                example: 3
              gender:
                type: integer
                example: 1
              weight:
                type: number
                format: float
                example: 25.5
              description:
                type: string
                example: "Friendly and energetic"
              created_at:
                type: string
                format: date-time
                example: "2023-06-01T14:00:00Z"
              updated_at:
                type: string
                format: date-time
                example: "2023-06-01T14:00:00Z"

    # 寄养服务相关
...
    BoardingServiceResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            service_id:
              type: string
              example: "s123456789"
            provider_id:
              type: string
              example: "u123456789"
            provider_name:
              type: string
              example: "John Doe"
            title:
              type: string
              example: "Luxury Pet Boarding"
            price:
              type: number
              format: float
              example: 50.00
            location:
              type: string
              example: "123 Main St, City"
            pet_type:
              type: string
              example: "Dog"
            capacity:
              type: integer
              example: 5
            available:
              type: integer
              example: 1
            rating:
              type: number
              format: float
              example: 4.8
            review_count:
              type: integer
              example: 25
            description:
              type: string
              example: "Luxury boarding with premium care"
            facilities:
              type: array
              items:
                type: string
              example: ["Air Conditioning", "Heated Floors", "Outdoor Play Area"]
            services:
              type: array
              items:
                type: string
              example: ["Daily Walks", "Grooming", "Training"]
            rules:
              type: string
              example: "Pets must be vaccinated and friendly"
            photos:
              type: array
              items:
                type: string
              example: ["url1.jpg", "url2.jpg"]
            created_at:
              type: string
              format: date-time
              example: "2023-06-01T17:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-06-01T17:00:00Z"
    BoardingServiceUpdateRequest:
      type: object
      properties:
        title:
          type: string
          example: "Premium Luxury Pet Boarding"
        price:
          type: number
          format: float
          example: 55.00
        capacity:
          type: integer
          example: 6
        available:
          type: integer
          enum: [0, 1]
          example: 0
        description:
          type: string
          example: "Updated description"
...
    ReviewCreateRequest:
      type: object
      required:
        - order_id
        - rating
      properties:
        order_id:
          type: string
          example: "o123456789"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        content:
          type: string
          example: "Excellent service, my pet was very happy"
        photos:
          type: array
          items:
            type: string
          example: ["url1.jpg", "url2.jpg"]
    ReviewResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            review_id:
              type: string
              example: "r123456789"
            order_id:
              type: string
              example: "o123456789"
            user_id:
              type: string
              example: "u123456789"
            user_name:
              type: string
              example: "John Doe"
            service_id:
              type: string
              example: "s123456789"
            service_title:
              type: string
              example: "Luxury Pet Boarding"
            rating:
              type: integer
              example: 5
            content:
              type: string
              example: "Excellent service, my pet was very happy"
            photos:
              type: array
              items:
                type: string
              example: ["url1.jpg", "url2.jpg"]
            created_at:
              type: string
              format: date-time
              example: "2023-06-01T23:00:00Z"
            updated_at:
              type: string
              format: date-time
              example: "2023-06-01T23:00:00Z"
    ReviewListResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        data:
          type: object
          properties:
            total:
              type: integer
              example: 30
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            average_rating:
              type: number
              format: float
              example: 4.8
            reviews:
              type: array
              items:
                type: object
                properties:
                  review_id:
                    type: string
                    example: "r123456789"
                  user_id:
                    type: string
                    example: "u123456789"
                  user_name:
                    type: string
                    example: "John Doe"
                  service_id:
                    type: string
                    example: "s123456789"
                  service_title:
                    type: string
                    example: "Luxury Pet Boarding"
                  rating:
                    type: integer
                    example: 5
                  content:
                    type: string
                    example: "Excellent service, my pet was very happy"
                  photos:
                    type: array
                    items:
                      type: string
                    example: ["url1.jpg", "url2.jpg"]
                  created_at:
                    type: string
                    format: date-time
                    example: "2023-06-01T23:00:00Z"

    # 通知相关
...